core functionality PostgresLinker -> PostgresAPI
